+++ sakura_core/charset/CCodeBase.h
@@ -61,14 +61,14 @@ public:
 	virtual void GetEol(CMemory* pcmemEol, EEolType eEolType){ S_GetEol(pcmemEol,eEolType); }	//!< 改行データ取得
 
 	// 文字コード表示用		2008/6/9 Uchi
-	virtual EConvertResult UnicodeToHex(const wchar_t* cSrc, const int iSLen, WCHAR* pDst, const CommonSetting_Statusbar* psStatusbar);			//!< UNICODE → Hex 変換
+	virtual EConvertResult UnicodeToHex(const wchar_t* cSrc, const SSIZE_T iSLen, WCHAR* pDst, const CommonSetting_Statusbar* psStatusbar);			//!< UNICODE → Hex 変換
 
 	// 変換エラー処理（１バイト <-> U+D800 から U+D8FF）
-	static int BinToText(const unsigned char *pSrc, const int nLen, unsigned short *pDst);
+	static int BinToText(const unsigned char *pSrc, const SSIZE_T nLen, unsigned short *pDst);
 	static int TextToBin(const unsigned short cSrc);
 
 	// MIME Header デコーダ
-	static bool MIMEHeaderDecode(const char* pSrc, const int nSrcLen, CMemory* pcMem, const ECodeType eCodetype);
+	static bool MIMEHeaderDecode(const char* pSrc, const SSIZE_T nSrcLen, CMemory* pcMem, const ECodeType eCodetype);
 
 	// CShiftJisより移動 2010/6/13 Uchi
 	static void S_GetEol(CMemory* pcmemEol, EEolType eEolType);	//!< 改行データ取得
@@ -77,9 +77,9 @@ public:
 /*!
 	バイナリ１バイトを U+DC00 から U+DCFF までに対応付ける
 */
-inline int CCodeBase::BinToText( const unsigned char *pSrc, const int nLen, unsigned short *pDst )
+inline int CCodeBase::BinToText( const unsigned char *pSrc, const SSIZE_T nLen, unsigned short *pDst )
 {
-	int i;
+	SSIZE_T i;
 
 	for( i = 0; i < nLen; ++i ){
 		pDst[i] = static_cast<unsigned short>(pSrc[i]) + 0xdc00;
